requirements.txt
# fastapi==0.115.0 
# uvicorn==0.30.6 
# python-dotenv==1.0.1
# langchain
# langchain_huggingface
# sentence-transformers==3.1.1 
# faiss-cpu==1.8.0 
# pdfplumber==0.11.4 
# transformers==4.44.2 
# # #torch==2.4.1
fastapi[standard]
uvicorn[standard]
langchain-community
langchain-huggingface
sentence-transformers
transformers
torch
faiss-cpu
pdfplumber
python-dotenv
numpy
hf_xet
pytest